feat: Add fabman access button
This commit adds the `Finish training` button below the laser cutter
video. After a user has started the video above this button, the button
is unlocked and the user is forwarded to the makerman access granting
service. This behaviour can be copied for other trainings as well, by
copying the button and adjusting the training id in the link.