Fabman plotter (#20)
* fix: Hardware Studio page responsive design

The images on the page caused the mobile view to break and extend over
the screen borders of the device. Now the images will scale according to
the display.

* fix: remove Creality printers

As we don't have any creality printers anymore we can safely remove the
information for them from the website.

* feat: Add Fabman button to plotter

As the plotter now also has a Fabman bridge, we want to link to the
Fabman account creation again.